backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-cloud-bigquery==3.20.1
google-auth==2.29.0
pydantic==2.7.1
python-dotenv==1.0.1
aiofiles==23.2.1
